Time: 0 ms (100%), Space: 9.4 MB (40.04%) - LeetHub

+++ 0940-distinct-subsequences-ii/0940-distinct-subsequences-ii.cpp
@@ -4,7 +4,7 @@ public:
     int distinctSubseqII(string s) {
         int n = s.length();
         vector<int> last(26, 0);
-        vector<long> dp(n+1, 0);
+        vector<uint32_t> dp(n+1, 0);
         dp[0] = 0;
         for (int i = 1; i <= n; i++) {
             int j = last[s[i-1]-'a'];
